Return a list of stripped paths
This adds a #validate method to the LogValidator class which returns a list of
validated paths from the supplied logfile. Since we know we will only need to
count the number of paths visited, we strip off any additional data supplied,
such as the IP address.

I've chosen to use File#each_line in this case as this prevents the need to
load the whole file into memory and instead processes each line sequentially.