Modify LogValidator to return path and IP address
In order to faciliate counting unique visits to pages, we adapt the
LogValidator class to return paths and IP addresses. This includes some light
validation of the IP address to exclude those with character that are not
digits or full stops.

We also modify the PageVisitCounter to accept a nested array of path data
instead of the original flat array of paths.